+++ src/main/java/com/nightingale/simplelearning/dao/impl/TestDAOImpl.java
@@ -1,11 +1,9 @@
 package com.nightingale.simplelearning.dao.impl;
 
-import com.nightingale.simplelearning.dao.CourseDAO;
 import com.nightingale.simplelearning.dao.QuestionDAO;
+import com.nightingale.simplelearning.dao.ResultDAO;
 import com.nightingale.simplelearning.dao.TestDAO;
-import com.nightingale.simplelearning.dao.mapper.CourseListRowMapper;
 import com.nightingale.simplelearning.dao.mapper.TestRowMapper;
-import com.nightingale.simplelearning.model.Course;
 import com.nightingale.simplelearning.model.Question;
 import com.nightingale.simplelearning.model.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -30,13 +28,10 @@ public class TestDAOImpl implements TestDAO {
     private QuestionDAO questionDAO;
 
     @Autowired
+    private ResultDAO resultDAO;
+
+    @Autowired
     private TestRowMapper testRowMapper;
-
-    @Autowired
-    private CourseDAO courseDAO;
-
-    @Autowired
-    private CourseListRowMapper courseListRowMapper;
 
     private static final Logger LOGGER = Logger.getLogger(TestDAOImpl.class.getName());
 
@@ -44,8 +39,10 @@ public class TestDAOImpl implements TestDAO {
     public Test getTestById(BigInteger id) {
         try {
             Test test = jdbcTemplate.queryForObject(SELECT_TEST_BY_ID, testRowMapper, id);
-            List<Question> questions = questionDAO.getAllQuestionsByTestId(id);
-            test.setQuestions(questions);
+            if (test != null) {
+                List<Question> questions = questionDAO.getAllQuestionsByTestId(id);
+                test.setQuestions(questions);
+            }
             return test;
         } catch (DataAccessException dataAccessException) {
             LOGGER.log(Level.WARNING, dataAccessException.getMessage(), dataAccessException);
@@ -57,6 +54,21 @@ public class TestDAOImpl implements TestDAO {
     public List<Test> getAllTestsByCourseId(BigInteger courseId) {
         try {
             return jdbcTemplate.query(SELECT_TESTS_BY_COURSE_ID, testRowMapper, courseId);
+        } catch (DataAccessException dataAccessException) {
+            LOGGER.log(Level.WARNING, dataAccessException.getMessage(), dataAccessException);
+            return null;
+        }
+    }
+
+    @Override
+    public Test getTestByResultId(BigInteger id) {
+        try {
+            Test test = jdbcTemplate.queryForObject(SELECT_TEST_BY_RESULT_ID, testRowMapper, id);
+            if (test!=null) {
+                List<Question> questions = questionDAO.getAllQuestionsByTestId(BigInteger.valueOf(test.getTestId()));
+                test.setQuestions(questions);
+            }
+            return test;
         } catch (DataAccessException dataAccessException) {
             LOGGER.log(Level.WARNING, dataAccessException.getMessage(), dataAccessException);
             return null;
@@ -77,14 +89,8 @@ public class TestDAOImpl implements TestDAO {
     @Transactional(rollbackFor = DataAccessException.class)
     public boolean deleteAllTestsByCourseId(BigInteger id) {
         try {
-            Course course = jdbcTemplate.queryForObject(courseDAO.SELECT_COURSE_BY_ID, courseListRowMapper, id);
-            if (course!=null) {
-                jdbcTemplate.update(DELETE_TESTS_BY_COURSE_ID, id);
-                return true;
-            } else {
-                LOGGER.log(Level.WARNING, "No Course with given ID");
-                return false;
-            }
+            jdbcTemplate.update(DELETE_TESTS_BY_COURSE_ID, id);
+            return true;
         } catch (DataAccessException dataAccessException) {
             LOGGER.log(Level.WARNING, dataAccessException.getMessage(), dataAccessException);
             TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
@@ -113,12 +119,17 @@ public class TestDAOImpl implements TestDAO {
     @Transactional(rollbackFor = DataAccessException.class)
     public boolean delete(BigInteger id) {
         try {
-            //TRY TO DELETE TEST'S QUESTIONS FIRST
-            boolean questionDeletionSuccessfulResult = questionDAO.deleteAllQuestionsByTestId(id);
-            //IF THAT'S DONE, TRY TO DELETE TEST ITSELF
-            if (questionDeletionSuccessfulResult) {
-                jdbcTemplate.update(DELETE_TEST_BY_ID, id);
-                return true;
+            Test test = getTestById(id);
+            if (test != null) {
+                //TRY TO DELETE TEST'S QUESTIONS AND RESULTS FIRST
+                boolean questionDeletionSuccessfulResult = questionDAO.deleteAllQuestionsByTestId(id);
+                boolean resultsDeletionSuccessfulResult = resultDAO.deleteAllResultsByTestId(id);
+                //IF THAT'S DONE, TRY TO DELETE TEST ITSELF
+                if (questionDeletionSuccessfulResult && resultsDeletionSuccessfulResult) {
+                    jdbcTemplate.update(DELETE_TEST_BY_ID, id);
+                    return true;
+                } else
+                    return false;
             } else
                 return false;
         } catch (DataAccessException dataAccessException) {
@@ -136,7 +147,7 @@ public class TestDAOImpl implements TestDAO {
             return false;
         }
         try {
-            Test test = jdbcTemplate.queryForObject(SELECT_TEST_BY_ID, testRowMapper, id);
+            Test test = getTestById(id);
             if (test!=null) {
                 jdbcTemplate.update(UPDATE_TEST_BY_ID, newTest.getTitle(), newTest.getTime(), id);
                 return true;
